django/requirements.txt
appdirs==1.4.4
asgiref==3.3.4
black==21.5b0
click==7.1.2
Django==3.2.2
django-js-asset==1.2.2
django-mptt==0.12.0
djangorestframework==3.12.4
isort==5.8.0
mypy-extensions==0.4.3
pathspec==0.8.1
Pillow==8.2.0
pytz==2021.1
regex==2021.4.4
sqlparse==0.4.1
toml==0.10.2
